Edit profile form
edit profile form will allow users to edit their profile with the system
ideally, the defaults will be what the user used to register on sign up. about me section will be blank and the user can edit it here.
the other fields are not required as it is assumed that the user will only edit the form when need be and is not a requirement